Add support for auto-configuring SimpleMessageListenerContainer
This commit introduces a `SimpleJmsListenerContainerFactoryConfigurer`
that can be used to conveniently apply settings provided
by JMS auto-configuration to `SimpleJmsListenerContainerFactory`
instances.
