+++ module/spring-boot-jms/src/main/java/org/springframework/boot/jms/autoconfigure/SimpleJmsListenerContainerFactoryConfigurer.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2012-present the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.springframework.boot.jms.autoconfigure;
+
+import jakarta.jms.ConnectionFactory;
+
+import org.springframework.jms.config.SimpleJmsListenerContainerFactory;
+
+/**
+ * Configures {@link SimpleJmsListenerContainerFactory} with sensible defaults.
+ *
+ * @author Vedran Pavic
+ * @since 4.0.0
+ */
+public final class SimpleJmsListenerContainerFactoryConfigurer
+		extends AbstractJmsListenerContainerFactoryConfigurer<SimpleJmsListenerContainerFactory> {
+
+	@Override
+	protected void configure(SimpleJmsListenerContainerFactory factory, ConnectionFactory connectionFactory,
+			JmsProperties jmsProperties) {
+	}
+
+}
